+++ Examples/FTMobileSDKUnitTests/Helper/FTNetworkMock.m
@@ -17,7 +17,9 @@ static NSString *urlStr;
     urlStr = urlString;
 }
 + (void)registerHandler:(void (^)(void))handler{
-    g_handler = handler;
+    @synchronized (self) {
+        g_handler = handler;
+    }
 }
 + (id<OHHTTPStubsDescriptor>)networkOHHTTPStubs{
     return [OHHTTPStubs stubRequestsPassingTest:^BOOL(NSURLRequest *request) {
@@ -42,9 +44,11 @@ static NSString *urlStr;
         NSData *requestData = [data dataUsingEncoding:NSUTF8StringEncoding];
         if(g_handler){
             dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(0.1 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
-                if(g_handler){
-                    g_handler();
-                    g_handler = nil;
+                @synchronized (self) {
+                    if(g_handler){
+                        g_handler();
+                        g_handler = nil;
+                    }
                 }
             });
         }