+++ Examples/FTMobileSDKUnitTests/Resource/FTNetworkInterceptorTest.m
@@ -42,14 +42,16 @@
     [self initSDKEnableAutoTrace:enable traceInterceptor:nil];
 }
 - (void)initSDKEnableAutoTrace:(BOOL)enable resourcePropertyProvider:(ResourcePropertyProvider)resourcePropertyProvider{
-    [self initSDKEnableAutoTrace:enable resourcePropertyProvider:resourcePropertyProvider traceInterceptor:nil];    
+    [self initSDKEnableAutoTrace:enable resourcePropertyProvider:resourcePropertyProvider traceInterceptor:nil errorFilter:nil];
 }
 - (void)initSDKEnableAutoTrace:(BOOL)enable traceInterceptor:(TraceInterceptor)traceInterceptor{
-    [self initSDKEnableAutoTrace:enable resourcePropertyProvider:nil traceInterceptor:traceInterceptor];
+    [self initSDKEnableAutoTrace:enable resourcePropertyProvider:nil traceInterceptor:traceInterceptor errorFilter:nil];
 }
 - (void)initSDKEnableAutoTrace:(BOOL)enable
       resourcePropertyProvider:(ResourcePropertyProvider)resourcePropertyProvider
-              traceInterceptor:(TraceInterceptor)traceInterceptor{
+              traceInterceptor:(TraceInterceptor)traceInterceptor
+                   errorFilter:(SessionTaskErrorFilter)errorFilter
+{
     NSProcessInfo *processInfo = [NSProcessInfo processInfo];
     NSString *url = [processInfo environment][@"ACCESS_SERVER_URL"];
     NSString *appid = [processInfo environment][@"APP_ID"];
@@ -59,6 +61,9 @@
     FTRumConfig *rumConfig = [[FTRumConfig alloc]initWithAppid:appid];
     if(resourcePropertyProvider){
         rumConfig.resourcePropertyProvider = resourcePropertyProvider;
+    }
+    if (errorFilter) {
+        rumConfig.sessionTaskErrorFilter = errorFilter;
     }
     rumConfig.enableTraceUserResource = enable;
     FTTraceConfig *traceConfig = [[FTTraceConfig alloc]init];
@@ -74,6 +79,68 @@
     [[FTTrackerEventDBTool sharedManger] deleteAllDatas];
 }
 #pragma mark - RUM
+- (void)testResourceLocalErrorFilter_session{
+    [self resourceLocalErrorFilter:YES enableGlobal:NO];
+}
+- (void)testResourceLocalErrorFilter_global{
+    [self resourceLocalErrorFilter:NO enableGlobal:YES];
+}
+- (void)testResourceLocalErrorFilter_priority{
+    [self resourceLocalErrorFilter:YES enableGlobal:YES];
+}
+- (void)resourceLocalErrorFilter:(BOOL)enableSession enableGlobal:(BOOL)enableGlobal{
+    [self initSDKEnableAutoTrace:YES resourcePropertyProvider:nil traceInterceptor:nil errorFilter:enableGlobal?^BOOL(NSError * _Nonnull error) {
+        if (error.code == NSURLErrorBadURL) {
+            return YES;
+        }
+        return NO;
+    }:nil];
+    NSURL *url = [NSURL URLWithString:@"http://test.error-filter.com"];
+    id<OHHTTPStubsDescriptor> stubs = [OHHTTPStubs stubRequestsPassingTest:^BOOL(NSURLRequest *request) {
+        return [request.URL.host isEqualToString:url.host];
+    } withStubResponse:^OHHTTPStubsResponse*(NSURLRequest *request) {
+        NSError* notConnectedError = [NSError errorWithDomain:NSURLErrorDomain code:NSURLErrorCancelled userInfo:@{NSLocalizedDescriptionKey:@"An asynchronous load has been canceled."}];
+        return [OHHTTPStubsResponse responseWithError:notConnectedError];
+    }];
+    FTURLSessionDelegate *ftDelegate = [[FTURLSessionDelegate alloc]init];
+    if (enableSession) {
+        ftDelegate.errorFilter = ^BOOL(NSError * _Nonnull error) {
+            if (error.code == NSURLErrorCancelled) {
+                return YES;
+            }
+            return NO;
+        };
+    }
+    XCTestExpectation *expectation = [self expectationWithDescription:@"request"];
+
+    NSURLSession *session = [NSURLSession sessionWithConfiguration:[NSURLSessionConfiguration ephemeralSessionConfiguration] delegate:ftDelegate delegateQueue:nil];
+    
+    NSURLSessionDataTask *task = [session dataTaskWithURL:url completionHandler:^(NSData * _Nullable data, NSURLResponse * _Nullable response, NSError * _Nullable error) {
+        [expectation fulfill];
+    }];
+    [task resume];
+    
+    [self waitForExpectations:@[expectation]];
+    [NSThread sleepForTimeInterval:0.5];
+
+    [[FTGlobalRumManager sharedInstance].rumManager syncProcess];
+    NSArray *newArray = [[FTTrackerEventDBTool sharedManger] getFirstRecords:10 withType:FT_DATA_TYPE_RUM];
+    __block int hasResourceCount = 0, hasErrorCount = 0;
+    [FTModelHelper resolveModelArray:newArray callBack:^(NSString * _Nonnull source, NSDictionary * _Nonnull tags, NSDictionary * _Nonnull fields, BOOL * _Nonnull stop) {
+        if ([source isEqualToString:FT_RUM_SOURCE_RESOURCE]) {
+            hasResourceCount ++;
+        }else if ([source isEqualToString:FT_RUM_SOURCE_ERROR]){
+            hasErrorCount ++;
+        }
+    }];
+    XCTAssertTrue(hasResourceCount == 1);
+    if (enableSession) {
+        XCTAssertTrue(hasErrorCount == 0);
+    }else if (enableGlobal){
+        XCTAssertTrue(hasErrorCount == 1);
+    }
+    [OHHTTPStubs removeStub:stubs];
+}
 /**
  *  RumAutoTrace = NO
  *  Session.ResourcePropertyProvider != nil