+++ FTMobileSDK/FTMobileAgent/Core/FTGlobalRumManager.m
@@ -11,7 +11,7 @@
 #import "FTGlobalRumManager.h"
 #import "FTLog+Private.h"
 #if !TARGET_OS_TV
-#import "FTWKWebViewHandler.h"
+#import "FTWKWebViewHandler+Private.h"
 #import "FTWKWebViewJavascriptBridge.h"
 #endif
 #import "FTLongTaskManager.h"
@@ -81,7 +81,7 @@ static dispatch_once_t onceToken;
         [dependencies.writer lastFatalErrorIfFound:0];
     }
 #if !TARGET_OS_TV
-    [FTWKWebViewHandler sharedInstance].rumTrackDelegate = self;
+    [[FTWKWebViewHandler sharedInstance] startWithEnableTraceWebView:rumConfig.enableTraceWebView allowWebViewHost:rumConfig.allowWebViewHost rumDelegate:self];
 #endif
     [FTExternalDataManager sharedManager].delegate = self.rumManager;
 }
@@ -154,10 +154,10 @@ static dispatch_once_t onceToken;
 - (void)shutDown{
     [[FTAppLifeCycle sharedInstance] removeAppLifecycleDelegate:self];
     [[FTAutoTrackHandler sharedInstance] shutDown];
-#if !TARGET_OS_TV
-    [FTWKWebViewHandler sharedInstance].enableTrace = NO;
-#endif
     [_longTaskManager shutDown];
+#if !TARGET_OS_TV
+    [[FTWKWebViewHandler sharedInstance] shutDown];
+#endif
     onceToken = 0;
     sharedInstance = nil;
     FTInnerLogInfo(@"[RUM] SHUT DOWN");