+++ FTMobileSDK/FTMobileSDK.h
@@ -30,3 +30,5 @@ FOUNDATION_EXPORT const unsigned char FTMobileSDKVersionString[];
 #import <FTMobileSDK/FTLog.h>
 #import <FTMobileSDK/FTTraceContext.h>
 #import <FTMobileSDK/FTLoggerConfig.h>
+#import <FTMobileSDK/FTConstants.h>
+#import <FTMobileSDK/FTWKWebViewHandler.h>