+++ FTMobileSDK/FTSDKCore/FTWKWebView/WKWebView+FTAutoTrack.h
@@ -14,7 +14,6 @@ NS_ASSUME_NONNULL_BEGIN
 -(WKNavigation *)ft_loadRequest:(NSURLRequest *)request;
 -(WKNavigation *)ft_loadHTMLString:(NSString *)string baseURL:(NSURL *)baseURL;
 -(WKNavigation *)ft_loadFileURL:(NSURL *)URL allowingReadAccessToURL:(NSURL *)readAccessURL;
--(WKNavigation *)ft_reload;
 @end
 NS_ASSUME_NONNULL_END
 #endif