+++ FTMobileSDK/FTSDKCore/RemoteConfig/FTRemoteConfigManager.h
@@ -1,0 +1,27 @@
+//
+//  FTRemoteConfigManager.h
+//  FTMobileSDK
+//
+//  Created by hulilei on 2025/6/5.
+//  Copyright © 2025 DataFlux-cn. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+#import "FTRemoteConfigurationProtocol.h"
+NS_ASSUME_NONNULL_BEGIN
+
+@interface FTRemoteConfigManager : NSObject<FTRemoteConfigurationDataSource>
+
+@property (nonatomic, weak) id<FTRemoteConfigurationProtocol> delegate;
+
++ (instancetype)sharedInstance;
+
+- (void)enable:(BOOL)enable updateInterval:(int)updateInterval;
+/// 请求远程配置
+- (void)updateRemoteConfig;
+
+- (void)updateRemoteConfigWithMiniUpdateInterval:(int)miniUpdateInterval callback:(nullable void (^)(BOOL, NSDictionary<NSString *,id> * _Nullable))callback;
+- (void)shutDown;
+@end
+
+NS_ASSUME_NONNULL_END