+++ FTMobileSDK/FTSDKCore/include/FTSDKCore.h
@@ -44,6 +44,8 @@
 #import "FTTrackerEventDBTool.h"
 #import "FTWebViewJavascriptBridgeBase.h"
 #import "FTWKWebViewHandler.h"
+#import "FTWKWebViewHandler+Private.h"
 #import "FTWKWebViewJavascriptBridge.h"
 #import "FTDataWriterWorker.h"
 #import "FTLoggerConfig.h"
+#import "FTRemoteConfigManager.h"